Decouple geometry from SVG, introduce Renderer
In order to facilitate generating graphics in formats other than SVG,
the main geometry structs needed to be independent of SVG. Multiple
traits were added, such as SvgRenderable, Pathable, and Pointable.
This setup is far cleaner than it was before, but with this change,
I removed some methods that are unique to some types, that would be
better implemented as traits, since most geometry types implement
Pathable, and therefore can share a ton of code.